[Feat] 재화 - DB 연동 추가
