+++ Assets/08. LSW_Folder/Scripts/Game/GameSceneUIController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 3f76870f83ac2ca468d06cc261e8c434
