+++ Assets/08. LSW_Folder/Scripts/Game/GameUIController.cs
@@ -1,46 +1,23 @@
-using System.Collections.Generic;
-using TMPro;
 using UnityEngine;
-using UnityEngine.UI;
 
 public class GameUIController : UIController<GameUIController.GameUIType>
 {
-    [SerializeField] private List<Toggle> _toggleGroup;
-    [SerializeField] private TextMeshProUGUI _gold;
-    
     public enum GameUIType
     {
-        DungeonPanel,
-        ShopPanel
+        AttendancePanel,
     }
     
     private void Start()
     {
-        for (int i = 0; i < _toggleGroup.Count; i++)
-        {
-            int index = i;
-            _toggleGroup[index].onValueChanged.AddListener((isOn) =>
-            {
-                _uiList[index].gameObject.SetActive(isOn);
-            });
-        }
-        
         foreach (var ui in _uiList)
         {
-            if (ui is DungeonPanel dungeonPanel)
+            if (ui is AttendancePanel attendancePanel)
             {
-            }
-            
-            else if (ui is ShopPanel shopPanel)
-            {
+                attendancePanel.OnTouchedExitBtn += () =>
+                {
+                    HideUI(GameUIType.AttendancePanel);
+                };
             }
         }
-        DatabaseManager.Instance.DisplayCreditData();
-        DatabaseManager.Instance.OnChangedCreditData += SetText;
-    }
-
-    private void SetText()
-    {
-        _gold.text = $"{DatabaseManager.Instance.PlayerData.CreditData.Gold.ToString()} Gold";
     }
 }
