+++ Assets/08. LSW_Folder/Scripts/Game/GameUIController.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 3f76870f83ac2ca468d06cc261e8c434
+guid: ae9bd831e792ec9449b1d543f9c43a27
